+++ nodeserver/routes/scheduledTasksRoute.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const { getScheduledTasks } = require('../scheduler');
+const { getScheduledTasks } = require('../../nodebackend/build/scheduler');
 const logger = require('../../nodebackend/build/logger').default;
 
 router.get('/', async (req, res) => {